src/main/java/easy/No35SearchInsert.java
package easy;

/**
 * 给定一个排序数组和一个目标值，在数组中找到目标值，并返回其索引。如果目标值不存在于数组中，返回它将会被按顺序插入的位置。
 * 你可以假设数组中无重复元素。
 *
 * 示例 1:
 * 输入: [1,3,5,6], 5
 * 输出: 2
 *
 * 示例 2:
 * 输入: [1,3,5,6], 2
 * 输出: 1
 *
 * 示例 3:
 * 输入: [1,3,5,6], 7
 * 输出: 4
 *
 * 示例 4:
 * 输入: [1,3,5,6], 0
 * 输出: 0
 */
public class No35SearchInsert {
    public int searchInsert(int[] nums, int target) {
        int left = 0;
        int right = nums.length - 1;
        int needInsert = 0;
        while (left <= right){
            int media = left + (right - left) / 2;
            if (target == nums[media]){
                return media;
            }else if (nums[media] < target){
                left = media + 1;
                needInsert = media + 1;
            }else {
                right = media - 1;
                needInsert = media;
            }
        }
        return needInsert;
    }
}
